had to put music stuff in gamemanager
